test(Lotto): modify lotto test
Lotto test 파일이 두 개인 것을 하나로 합쳤습니다.
